pyproject.toml
[tool.poetry]
name = "webhook-listener"
version = "0.1.0"
description = "A simple webhook listener for ngrok events"
authors = ["openhands <openhands@all-hands.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
python-multipart = "^0.0.9"  # For handling form data
requests = "^2.0"
python-dotenv = "^1.0"
jinja2 = "^3.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"